Move most of the button stuff to a separate file
